DatabaseManagerを削除し、その役割はMemoManagerにまとめた
